Fix: ユーザー登録機能をAccountControllerに移動。
