--- src/main/java/com/example/raha/controller/LoginController.java
@@ -1,11 +1,0 @@
-package com.example.raha.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequiredArgsConstructor
-public class LoginController {
-
-}
